fix(reload): refuse a hot reload that routes to antigravity without a credential
Move `antigravity_migration_error` and its routing helpers
(`routes_to_antigravity`, `routes_to_antigravity_cli`,
`routes_to_kind`, `warn_if_routes_to_antigravity_cli`) from the
binary crate (main.rs) into the lib crate
(auth::antigravity), so reload::reload can reuse the same boot
guard main.rs's serve() already applies.

The check runs before Config::load's candidate ever touches the
live ArcSwap state, matching reload's documented fail-safe
contract: a config edit that would newly route to the native
antigravity upstream with no credential file is refused, and the
previously-running config keeps serving.

Adds ConfigError::AntigravityMigrationRequired and two reload.rs
tests (refused without a credential, keeps the previous state;
succeeds with one present).
