+++ src/main.rs
@@ -5,6 +5,10 @@ use std::sync::OnceLock;
 use anyhow::Context;
 use clap::{Parser, Subcommand};
 use shunt::{
+    auth::antigravity::{
+        antigravity_migration_error, routes_to_antigravity, routes_to_antigravity_cli,
+        warn_if_routes_to_antigravity_cli,
+    },
     blueprints::{self, AddKind},
     config::{Config, OtelConfig, SentryConfig},
     init, server,
@@ -707,73 +711,6 @@ fn init_telemetry(config: Option<&OtelConfig>) -> Option<TelemetryGuard> {
     }
 }
 
-/// Whether any routing surface can actually reach a provider of `kind`.
-///
-/// Presence in the provider map is not routing: `Config::default()` seeds
-/// built-in `antigravity` and `antigravity-cli` entries, so a config whose
-/// traffic goes entirely to Anthropic still "has" them. Warming on that spawned
-/// a ~20s `agy models` subprocess on every startup with `agy` installed, for
-/// nothing.
-///
-/// Checks every way a request can select a provider: the default provider,
-/// exact routes, prefix routes, and a model's `upstream_model` map.
-fn routes_to_kind(config: &Config, kind: shunt::config::ProviderKind) -> bool {
-    let is_kind = |name: &str| {
-        config
-            .providers
-            .get(name)
-            .is_some_and(|provider| provider.kind == kind)
-    };
-
-    is_kind(&config.server.default_provider)
-        || config.routes.iter().any(|route| is_kind(&route.provider))
-        || config
-            .route_prefixes
-            .iter()
-            .any(|prefix| is_kind(&prefix.provider))
-        || config.models.iter().any(|model| {
-            model
-                .upstream_model
-                .as_ref()
-                .is_some_and(|map| map.keys().any(|provider| is_kind(provider)))
-        })
-}
-
-/// Whether the deprecated `agy` subprocess transport is reachable.
-fn routes_to_antigravity_cli(config: &Config) -> bool {
-    routes_to_kind(config, shunt::config::ProviderKind::AntigravityCli)
-}
-
-/// Whether the native Antigravity HTTP upstream is reachable.
-fn routes_to_antigravity(config: &Config) -> bool {
-    routes_to_kind(config, shunt::config::ProviderKind::Antigravity)
-}
-
-/// Issue #368 Stage 1: warn when config still routes to the deprecated `agy`
-/// subprocess transport. Split from the call site so the warning is testable
-/// without spinning up `serve()`.
-fn warn_if_routes_to_antigravity_cli(config: &Config) {
-    if routes_to_antigravity_cli(config) {
-        tracing::warn!(
-            "config routes to `kind = \"antigravity_cli\"`, which is deprecated. \
-             Switch to `kind = \"antigravity\"` for the native HTTP transport."
-        );
-    }
-}
-
-/// The startup refusal for a routed native Antigravity provider with no
-/// credential. Split from the filesystem probe so the message is testable
-/// without depending on what happens to exist in the test environment's home
-/// directory.
-fn antigravity_migration_error(credential_exists: bool) -> Option<String> {
-    (!credential_exists).then(|| {
-        "provider `antigravity` is routed but has no credential. It is now the native HTTP \
-         upstream, not the local `agy` CLI. Run `shunt login antigravity` to authenticate it, \
-         or route to `antigravity-cli` to stay on the deprecated subprocess transport."
-            .to_string()
-    })
-}
-
 #[cfg(test)]
 mod tests {
     use std::io;
@@ -1225,186 +1162,5 @@ mod tests {
             .expect("serve returns before the test deadline")
             .expect("serve task join")
             .expect("graceful shutdown returns Ok once drained");
-    }
-}
-
-#[cfg(test)]
-mod warm_gate_tests {
-    use std::io;
-    use std::sync::{Arc, Mutex};
-
-    use shunt::config::{Config, ModelConfig, ProviderKind, RouteConfig, RoutePrefixConfig};
-
-    use super::routes_to_antigravity;
-
-    struct BufferWriter {
-        buffer: Arc<Mutex<Vec<u8>>>,
-    }
-
-    impl std::io::Write for BufferWriter {
-        fn write(&mut self, bytes: &[u8]) -> io::Result<usize> {
-            self.buffer.lock().unwrap().extend_from_slice(bytes);
-            Ok(bytes.len())
-        }
-
-        fn flush(&mut self) -> io::Result<()> {
-            Ok(())
-        }
-    }
-
-    /// Turn the built-in `antigravity` provider into the one under test without
-    /// hand-building a provider: `Config::default()` already seeds it.
-    fn base() -> Config {
-        let config = Config::default();
-        assert_eq!(
-            config
-                .providers
-                .get("antigravity")
-                .map(|provider| provider.kind),
-            Some(ProviderKind::Antigravity),
-            "this test rests on the default config seeding an antigravity provider"
-        );
-        config
-    }
-
-    #[test]
-    fn a_routed_native_antigravity_without_a_credential_refuses_to_start() {
-        // Loud beats silent: routing to `antigravity` used to mean "run the
-        // local `agy` binary", so booting green on the HTTP transport instead
-        // would change credentials and egress underneath the operator.
-        let message = super::antigravity_migration_error(false)
-            .expect("a missing credential must refuse the boot");
-        assert!(message.contains("shunt login antigravity"), "{message}");
-        // Both ways forward have to be named, including staying on the old
-        // transport.
-        assert!(message.contains("antigravity-cli"), "{message}");
-    }
-
-    #[test]
-    fn a_present_credential_starts_normally() {
-        assert_eq!(super::antigravity_migration_error(true), None);
-    }
-
-    #[test]
-    fn the_two_antigravity_transports_are_detected_separately() {
-        // Routing to one must not warm or gate the other: the CLI path spawns a
-        // ~20s subprocess and the native path starts a version refresher.
-        let mut config = base();
-        config.server.default_provider = "antigravity".to_string();
-        assert!(super::routes_to_antigravity(&config));
-        assert!(!super::routes_to_antigravity_cli(&config));
-
-        let mut config = base();
-        config.server.default_provider = "antigravity-cli".to_string();
-        assert!(super::routes_to_antigravity_cli(&config));
-        assert!(!super::routes_to_antigravity(&config));
-    }
-
-    #[test]
-    fn a_routed_cli_config_warns_about_the_deprecated_kind() {
-        let output = Arc::new(Mutex::new(Vec::new()));
-        let writer_output = Arc::clone(&output);
-        let subscriber = tracing_subscriber::fmt()
-            .with_writer(move || BufferWriter {
-                buffer: Arc::clone(&writer_output),
-            })
-            .with_ansi(false)
-            .without_time()
-            .finish();
-
-        let mut config = base();
-        config.server.default_provider = "antigravity-cli".to_string();
-        tracing::subscriber::with_default(subscriber, || {
-            super::warn_if_routes_to_antigravity_cli(&config);
-        });
-        let logs = String::from_utf8(output.lock().unwrap().clone()).unwrap();
-
-        assert!(logs.contains("antigravity_cli"), "{logs}");
-        assert!(logs.contains("deprecated"), "{logs}");
-    }
-
-    #[test]
-    fn a_default_config_does_not_warn_about_the_deprecated_kind() {
-        let output = Arc::new(Mutex::new(Vec::new()));
-        let writer_output = Arc::clone(&output);
-        let subscriber = tracing_subscriber::fmt()
-            .with_writer(move || BufferWriter {
-                buffer: Arc::clone(&writer_output),
-            })
-            .with_ansi(false)
-            .without_time()
-            .finish();
-
-        tracing::subscriber::with_default(subscriber, || {
-            super::warn_if_routes_to_antigravity_cli(&base());
-        });
-        let logs = String::from_utf8(output.lock().unwrap().clone()).unwrap();
-
-        assert!(logs.is_empty(), "{logs}");
-    }
-
-    #[test]
-    fn default_config_does_not_route_to_antigravity() {
-        // The provider exists but nothing selects it. This is the whole point:
-        // presence is not routing.
-        assert!(!routes_to_antigravity(&base()));
-    }
-
-    #[test]
-    fn default_provider_pointing_at_antigravity_counts() {
-        let mut config = base();
-        config.server.default_provider = "antigravity".to_string();
-        assert!(routes_to_antigravity(&config));
-    }
-
-    #[test]
-    fn an_exact_route_counts() {
-        let mut config = base();
-        config.routes.push(RouteConfig {
-            model: "gemini-3.1-pro".to_string(),
-            provider: "antigravity".to_string(),
-            upstream_model: None,
-            effort: None,
-            service_tier: None,
-        });
-        assert!(routes_to_antigravity(&config));
-    }
-
-    #[test]
-    fn a_prefix_route_counts() {
-        let mut config = base();
-        config.route_prefixes.push(RoutePrefixConfig {
-            prefix: "gemini-".to_string(),
-            provider: "antigravity".to_string(),
-        });
-        assert!(routes_to_antigravity(&config));
-    }
-
-    #[test]
-    fn a_model_upstream_map_counts() {
-        let mut config = base();
-        config.models.push(ModelConfig {
-            id: "claude-gemini-via-agy".to_string(),
-            display_name: None,
-            upstream_model: Some(
-                [("antigravity".to_string(), "gemini-3.1-pro".to_string())]
-                    .into_iter()
-                    .collect(),
-            ),
-        });
-        assert!(routes_to_antigravity(&config));
-    }
-
-    #[test]
-    fn routing_to_a_non_antigravity_provider_does_not_count() {
-        let mut config = base();
-        config.routes.push(RouteConfig {
-            model: "some-model".to_string(),
-            provider: "anthropic".to_string(),
-            upstream_model: None,
-            effort: None,
-            service_tier: None,
-        });
-        assert!(!routes_to_antigravity(&config));
     }
 }
